==> R/3d_Analysis_RSF_Optimize.R <==
###################### RESEARCH LAB POWERPOINT - 2016.04.28  ###################

########################### LOAD PACKAGES AND DATA  ############################

suppressPackageStartupMessages(library(AICcmodavg))
suppressPackageStartupMessages(library(plyr))
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(ggmap))
suppressPackageStartupMessages(library(ggplot2))
suppressPackageStartupMessages(library(optimx))
suppressPackageStartupMessages(library(raster))
suppressPackageStartupMessages(library(rgdal))
suppressPackageStartupMessages(library(smoothie))
source('C:/Work/R/Functions/all.R')

####################  IMPORT POINT DATA AND RASTERS ############################

# Load point data
load(file=file.path(getwd(), "ua_data.RData"))

table(ua_data$id)

# Import rasters
base <- raster("C:/ArcGIS/Data/BlankRaster/maine_30mc.tif")
lc_30mc <- raster("C:/ArcGIS/Data/Landcover/Landcover/lc_30mc.tif")

################## CONVERT POINT DATA AND CROP LANDSCAPE RASTER ################

ua_data_sp <- SpatialPointsDataFrame(ua_data[,c("x","y")], ua_data,
  proj4string = CRS("+proj=utm +zone=19 ellps=WGS84"), bbox = NULL)
lc_crop <- crop(lc_30mc, extend(extent(ua_data_sp), 500))

###################### CONVERT LANDSCAPE RASTER ################################

# Convert all non-zero values to 1's

developed <- subs(lc_crop, data.frame(id=c(21:24), v=rep(1, 4))) ## Developed
developed[is.na(developed)] <- 0

forest <- subs(lc_crop, data.frame(id=c(41:43), v=rep(1, 3))) ## Forest
forest[is.na(forest)] <- 0

open_water <- subs(lc_crop, data.frame(id=11, v=1)) ## Open Water
open_water[is.na(open_water)] <- 0

pasture <- subs(lc_crop, data.frame(id=c(81, 82), v=rep(1, 2))) ## Pasture/Crop
pasture[is.na(pasture)] <- 0

shrub_herb <- subs(lc_crop, data.frame(id=c(52, 71), v=rep(1, 2))) ## Shrub/Herb
shrub_herb[is.na(shrub_herb)] <- 0

wetland <- subs(lc_crop, data.frame(id=c(90, 95), v=rep(1, 2))) ## Wetland
wetland[is.na(wetland)] <- 0



################ OPTIMIZATION PROCEDURE OF KERNEL BANDWIDTH ####################

load(file=file.path(getwd(), "all_mods.RData"))
load(file=file.path(getwd(), "opt_mods.RData"))
load(file=file.path(getwd(), "ua_data_lc_data.RData"))

# Likelihood function
# z = a vector of zeros and ones corresponding to available and used points
# cell = a vector of the cell numbers of each point
# surface = a data frame with x/y coordinates and values of the raster surface

  r1 <- rasterFromXYZ(surface1)

  r2 <- rasterFromXYZ(surface2)# surface is a df with X|Y|VALUES
#  zmat2 <- as.matrix(r2)

nll_kern_bw <- function(parms=parms, z=z, cell=cell, zmat1=zmat1, zmat2=zmat2){
  # Parameters
  b0 <- parms[1]         # intercept
  bx1 <- parms[2]         # slope of the relationship
  bwx1 <-  exp(parms[3])+1
  bx2 <- parms[4]# bw/spatial scale parameter
  bwx2 <-  exp(parms[5])+1
  # Create matrix
  zmat1 <- as.matrix(r1)
  zmat2 <- as.matrix(r1)
  # Gaussian-weighted smooth of matrix
  f <- kernel2dsmooth(zmat1,kernel.type="gauss", nx=nrow(r), ny=ncol(r),sigma=bwx1)
  values(r1) <- f
  # Gaussian-weighted smooth of matrix
  f2 <- kernel2dsmooth(zmat2,kernel.type="gauss", nx=nrow(r), ny=ncol(r),sigma=bwx2)
  values(r2) <- f2
  # Compute binomial success probabilities
  probs <- plogis(b0 + bx1*(r[cell])+ bx2(r[cell])
  # Evaluate log of binomial pmf
  tmp <- dbinom(z, 1, probs, log=TRUE)
  ll <- -1*sum(tmp)
  return(ll)
}

opt_mods_coef <- dplyr::left_join(opt_mods, all_mods, by = c("lc_type", "bw"))

#  for (i in 1:length(lc_types)){
#   test <- eval(parse(text = lc_types[i]))
#   print(test)
#  }

# Parameters
developed_table = opt_mods_coef %>% filter(lc_type == "developed")
parms_developed <- c(developed_table$intercept, developed_table$slope, log(developed_table$opt_bw/15))

forest_table = opt_mods_coef %>% filter(lc_type == "forest")
parms_forest <- c(forest_table$intercept, forest_table$slope, log(forest_table$opt_bw/15))

open_water_table = opt_mods_coef %>% filter(lc_type == "open_water")
parms_open_water <- c(open_water_table$intercept, open_water_table$slope, log(open_water_table$opt_bw/15))

pasture_table = opt_mods_coef %>% filter(lc_type == "pasture")
parms_pasture <- c(pasture_table$intercept, pasture_table$slope, log(pasture_table$opt_bw/15))

shrub_herb_table = opt_mods_coef %>% filter(lc_type == "shrub_herb")
parms_shrub_herb <- c(shrub_herb_table$intercept, shrub_herb_table$slope, log(shrub_herb_table$opt_bw/15))

wetland_table = opt_mods_coef %>% filter(lc_type == "wetland")
parms_wetland <- c(wetland_table$intercept, wetland_table$slope, log(wetland_table$opt_bw/15))

# Extract cell numbers
cell_developed <- extract(developed, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]
cell_forest <- extract(forest, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]
cell_open_water <- extract(open_water, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]
cell_pasture <- extract(pasture, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]
cell_shrub_herb <- extract(shrub_herb, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]
cell_wetland <- extract(wetland, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]

# Convert raster to data frame
surface_developed <- data.frame(coordinates(developed), values(developed))
surface_forest <- data.frame(coordinates(forest), values(forest))
surface_open_water <- data.frame(coordinates(open_water), values(open_water))
surface_pasture <- data.frame(coordinates(pasture), values(pasture))
surface_shrub_herb <- data.frame(coordinates(shrub_herb), values(shrub_herb))
surface_wetland <- data.frame(coordinates(wetland), values(wetland))
# test: surface <- data.frame(coordinates(eval(parse(text = lc_types[i]))), values(eval(parse(text = lc_types[i]))))

# OPTIMIZE!!!!!!!!!!!!

system.time(out_developed <- optimx(par=parms_developed, fn=nll_kern_bw,
  z=ua_data$point, cell=cell_developed, surface=surface_developed, control = list(trace=3)))
save(out_developed, file = "out_developed.RData")

system.time(out_forest <- optimx(par=parms_forest, fn=nll_kern_bw,
  z=ua_data$point, cell=cell_forest, surface=surface_forest, control = list(trace=3)))
save(out_forest, file = "out_forest.RData")

system.time(out_open_water <- optimx(par=parms_open_water, fn=nll_kern_bw,
  z=ua_data$point, cell=cell_open_water, surface=surface_open_water, control = list(trace=3)))
save(out_open_water, file = "out_open_water.RData")

system.time(out_pasture <- optimx(par=parms_pasture, fn=nll_kern_bw,
  z=ua_data$point, cell=cell_pasture, surface=surface_pasture, control = list(trace=3)))
save(out_pasture, file = "out_pasture.RData")

system.time(out_shrub_herb <- optimx(par=parms_shrub_herb, fn=nll_kern_bw,
  z=ua_data$point, cell=cell_shrub_herb, surface=surface_shrub_herb, control = list(trace=3)))
save(out_shrub_herb, file = "out_shrub_herb.RData")

system.time(out_wetland <- optimx(par=parms_wetland, fn=nll_kern_bw,
  z=ua_data$point, cell=cell_wetland, surface=surface_wetland, control = list(trace=3)))
save(out_wetland, file = "out_wetland.RData")


# Load optimx results

load(file="out_developed.RData")
load(file="out_forest.RData")
load(file="out_open_water.RData")
load(file="out_pasture.RData")
load(file="out_shrub_herb.RData")
load(file="out_wetland.RData")

(exp(out_developed[1,3])+1)*15
(exp(out_forest[1,3])+1)*15
(exp(out_open_water[1,3])+1)*15
(exp(out_pasture[1,3])+1)*15
(exp(out_shrub_herb[1,3])+1)*15
(exp(out_wetland[1,3])+1)*15


(exp(out_developed[2,3])+1)*15
(exp(out_forest[2,3])+1)*15
(exp(out_open_water[2,3])+1)*15
(exp(out_pasture[2,3])+1)*15
(exp(out_shrub_herb[2,3])+1)*15
(exp(out_wetland[2,3])+1)*15

log(150)
exp(5.010635)


# FAILED
system.time(out <- optim(par=parms, fn=nll_kern_bw, method = "L-BFGS-B",
      lower = c(-Inf, -Inf, log(1)), upper = rep(Inf, Inf, log(500)),
      z=ua_data$point, cell=cell1, surface=surface, control = list(trace=3,
        REPORT=1)))
# FAILED
system.time(out <- nlm(nll_kern_bw,parms,print.level=2,hessian=TRUE))


################ MULTI LAYER OPTIMX ############################################

# Create surfaces and matrixes

r1 <- surface1
r2 <- surface2
zmat1 <- as.matrix(r1)
zmat2 <- as.matrix(r2)

# Likelihood function
# z = a vector of zeros and ones corresponding to available and used points
# cell = a vector of the cell numbers of each point
# surface = a data frame with x/y coordinates and values of the raster surface

nll_kern_bw <- function(parms=parms, z=z, cell=cell, r1=r1, r2=r2,
    zmat1=zmat1, zmat2=zmat2){
  # Parameters
  b0 <- parms[1]             # intercept
  beta1 <- parms[2]            # slope of the relationship, surface 1
  bw1 <-  exp(parms[3])+1   # bw/spatial scale parameter, surface 1
  beta2 <- parms[4]            # slope of the relationship, surface 2
  bw2 <-  exp(parms[5])+1   # bw/spatial scale parameter, surface 2
  # Gaussian-weighted smooth of matrix
  f1 <- kernel2dsmooth(zmat1, kernel.type="gauss", nx=nrow(r1), ny=ncol(r1), sigma=bw1)
  values(r1) <- f1
  # Gaussian-weighted smooth of matrix
  f2 <- kernel2dsmooth(zmat2, kernel.type="gauss", nx=nrow(r2), ny=ncol(r2), sigma=bw2)
  values(r2) <- f2
  # Compute binomial success probabilities
  probs <- plogis(b0 + beta1*(r1[cell]) + beta2(r2[cell])
  # Evaluate log of binomial pmf
  tmp <- dbinom(z, 1, probs, log=TRUE)
  ll <- -1*sum(tmp)
  return(ll)
}

for (i in 1:length(lc_types)){
  lc_type_i <- lc_types[i]
  ua_data_sub <- ua_data %>% dplyr::select(point, contains(lc_type_i))
  lc_mods <- list(
    "0"   = glm(ua_data_sub[,1]~ua_data_sub[,2], ua_data_sub, family="binomial"),
    "15"  = glm(ua_data_sub[,1]~ua_data_sub[,3], ua_data_sub, family="binomial"),
    "30"  = glm(ua_data_sub[,1]~ua_data_sub[,4], ua_data_sub, family="binomial"),
    "45"  = glm(ua_data_sub[,1]~ua_data_sub[,5], ua_data_sub, family="binomial"),
    "60"  = glm(ua_data_sub[,1]~ua_data_sub[,6], ua_data_sub, family="binomial"),
    "75"  = glm(ua_data_sub[,1]~ua_data_sub[,7], ua_data_sub, family="binomial"),
    "90"  = glm(ua_data_sub[,1]~ua_data_sub[,8], ua_data_sub, family="binomial"),
    "105"  = glm(ua_data_sub[,1]~ua_data_sub[,9], ua_data_sub, family="binomial"),
    "120"  = glm(ua_data_sub[,1]~ua_data_sub[,10], ua_data_sub, family="binomial"),
    "135"  = glm(ua_data_sub[,1]~ua_data_sub[,11], ua_data_sub, family="binomial"),
    "150"  = glm(ua_data_sub[,1]~ua_data_sub[,12], ua_data_sub, family="binomial"),
    "165"  = glm(ua_data_sub[,1]~ua_data_sub[,13], ua_data_sub, family="binomial"),
    "180"  = glm(ua_data_sub[,1]~ua_data_sub[,14], ua_data_sub, family="binomial"),
    "195"  = glm(ua_data_sub[,1]~ua_data_sub[,15], ua_data_sub, family="binomial"),
    "210"  = glm(ua_data_sub[,1]~ua_data_sub[,16], ua_data_sub, family="binomial"),
    "225"  = glm(ua_data_sub[,1]~ua_data_sub[,17], ua_data_sub, family="binomial"),
    "240"  = glm(ua_data_sub[,1]~ua_data_sub[,18], ua_data_sub, family="binomial"),
    "255"  = glm(ua_data_sub[,1]~ua_data_sub[,19], ua_data_sub, family="binomial"),
    "270"  = glm(ua_data_sub[,1]~ua_data_sub[,20], ua_data_sub, family="binomial"),
    "285"  = glm(ua_data_sub[,1]~ua_data_sub[,21], ua_data_sub, family="binomial"),
    "300"  = glm(ua_data_sub[,1]~ua_data_sub[,22], ua_data_sub, family="binomial"),
    "315"  = glm(ua_data_sub[,1]~ua_data_sub[,23], ua_data_sub, family="binomial"),
    "330"  = glm(ua_data_sub[,1]~ua_data_sub[,24], ua_data_sub, family="binomial"),
    "345"  = glm(ua_data_sub[,1]~ua_data_sub[,25], ua_data_sub, family="binomial"),
    "360"  = glm(ua_data_sub[,1]~ua_data_sub[,26], ua_data_sub, family="binomial"),
    "375"  = glm(ua_data_sub[,1]~ua_data_sub[,27], ua_data_sub, family="binomial"),
    "390"  = glm(ua_data_sub[,1]~ua_data_sub[,28], ua_data_sub, family="binomial"),
    "405"  = glm(ua_data_sub[,1]~ua_data_sub[,29], ua_data_sub, family="binomial"),
    "420"  = glm(ua_data_sub[,1]~ua_data_sub[,30], ua_data_sub, family="binomial"),
    "435"  = glm(ua_data_sub[,1]~ua_data_sub[,31], ua_data_sub, family="binomial"),
    "450"  = glm(ua_data_sub[,1]~ua_data_sub[,32], ua_data_sub, family="binomial"),
    "465"  = glm(ua_data_sub[,1]~ua_data_sub[,33], ua_data_sub, family="binomial"),
    "480"  = glm(ua_data_sub[,1]~ua_data_sub[,34], ua_data_sub, family="binomial"),
    "495"  = glm(ua_data_sub[,1]~ua_data_sub[,35], ua_data_sub, family="binomial"),
    "510"  = glm(ua_data_sub[,1]~ua_data_sub[,36], ua_data_sub, family="binomial"),
    "525"  = glm(ua_data_sub[,1]~ua_data_sub[,37], ua_data_sub, family="binomial"),
    "540"  = glm(ua_data_sub[,1]~ua_data_sub[,38], ua_data_sub, family="binomial"),
    "555"  = glm(ua_data_sub[,1]~ua_data_sub[,39], ua_data_sub, family="binomial"),
    "570"  = glm(ua_data_sub[,1]~ua_data_sub[,40], ua_data_sub, family="binomial"),
    "585"  = glm(ua_data_sub[,1]~ua_data_sub[,41], ua_data_sub, family="binomial"),
    "600"  = glm(ua_data_sub[,1]~ua_data_sub[,42], ua_data_sub, family="binomial"),
    "615"  = glm(ua_data_sub[,1]~ua_data_sub[,43], ua_data_sub, family="binomial"),
    "630"  = glm(ua_data_sub[,1]~ua_data_sub[,44], ua_data_sub, family="binomial"),
    "645"  = glm(ua_data_sub[,1]~ua_data_sub[,45], ua_data_sub, family="binomial"),
    "660"  = glm(ua_data_sub[,1]~ua_data_sub[,46], ua_data_sub, family="binomial"),
    "675"  = glm(ua_data_sub[,1]~ua_data_sub[,47], ua_data_sub, family="binomial"),
    "690"  = glm(ua_data_sub[,1]~ua_data_sub[,48], ua_data_sub, family="binomial"),
    "705"  = glm(ua_data_sub[,1]~ua_data_sub[,49], ua_data_sub, family="binomial"),
    "720"  = glm(ua_data_sub[,1]~ua_data_sub[,50], ua_data_sub, family="binomial"),
    "735"  = glm(ua_data_sub[,1]~ua_data_sub[,51], ua_data_sub, family="binomial"),
    "750"  = glm(ua_data_sub[,1]~ua_data_sub[,52], ua_data_sub, family="binomial"),
    "825"  = glm(ua_data_sub[,1]~ua_data_sub[,53], ua_data_sub, family="binomial"),
    "900"  = glm(ua_data_sub[,1]~ua_data_sub[,54], ua_data_sub, family="binomial"),
    "975"  = glm(ua_data_sub[,1]~ua_data_sub[,55], ua_data_sub, family="binomial"),
    "1050"  = glm(ua_data_sub[,1]~ua_data_sub[,56], ua_data_sub, family="binomial"),
    "1125"  = glm(ua_data_sub[,1]~ua_data_sub[,57], ua_data_sub, family="binomial"),
    "1200"  = glm(ua_data_sub[,1]~ua_data_sub[,58], ua_data_sub, family="binomial"),
    "1275"  = glm(ua_data_sub[,1]~ua_data_sub[,59], ua_data_sub, family="binomial"),
    "1350"  = glm(ua_data_sub[,1]~ua_data_sub[,60], ua_data_sub, family="binomial"),
    "1425"  = glm(ua_data_sub[,1]~ua_data_sub[,61], ua_data_sub, family="binomial"),
    "1500"  = glm(ua_data_sub[,1]~ua_data_sub[,62], ua_data_sub, family="binomial")
  )
  mod_table <- aictab(lc_mods, second.ord = FALSE)
  View(mod_table)
  optimum_bw <- as.numeric(as.character(mod_table[1,1]))
  g <- ggplot(mod_table, aes(x=as.numeric(as.character(Modnames)), y = AIC)) +
    geom_line(color="red") + geom_point() + xlab("Bandwidth") + theme_no_legend +
    ggtitle(lc_type_i) + geom_vline(xintercept = optimum_bw, color="blue")
  print(g)
  SaveGGPlot(file = paste0("AIC_BW_", lc_type_i, ".jpg"))
}

lc_mods <- list(
  "0"   = glm(point~open_water0, ua_data, family="binomial"),
  "15"  = glm(point~open_water15, data=ua_data, family="binomial"),
  "30"  = glm(point~open_water30, data=ua_data, family="binomial"),
  "45"  = glm(point~open_water45, data=ua_data, family="binomial")
)

names(lc_mods)[[1]] <- as.character(100)

lc_mods <- list(
  "0"   = glm(point~LC,    ua_data, family="binomial"),
  "15"  = glm(point~LC15,  ua_data, family="binomial"),
  "30"  = glm(point~LC30,  ua_data, family="binomial"),
  "45"  = glm(point~LC45,  ua_data, family="binomial"),
  "60"  = glm(point~LC60,  ua_data, family="binomial"),
  "75"  = glm(point~LC75,  ua_data, family="binomial"),
  "90"  = glm(point~LC90,  ua_data, family="binomial"),
  "105" = glm(point~LC105, ua_data, family="binomial"),
  "120" = glm(point~LC120, ua_data, family="binomial"),
  "135" = glm(point~LC135, ua_data, family="binomial"),
  "150" = glm(point~LC150, ua_data, family="binomial"),
  "165" = glm(point~LC165, ua_data, family="binomial"),
  "180" = glm(point~LC180, ua_data, family="binomial"),
  "195" = glm(point~LC195, ua_data, family="binomial"),
  "210" = glm(point~LC210, ua_data, family="binomial"),
  "225" = glm(point~LC225, ua_data, family="binomial"),
  "240" = glm(point~LC240, ua_data, family="binomial"),
  "255" = glm(point~LC255, ua_data, family="binomial"),
  "270" = glm(point~LC270, ua_data, family="binomial"),
  "285" = glm(point~LC285, ua_data, family="binomial"),
  "300" = glm(point~LC300, ua_data, family="binomial")
)



predict_range <- with(ua_data, data.frame(open_water15 = seq(from = 0, to = 1,
  length.out = 100)))



lc_mods <- list(
  "15"   = glm(point~open_water15,    ua_data, family="binomial")
  )

predicted <- cbind(predict_range, predict(lc_mods[["15"]],
  newdata=predict_range, type="link", se=TRUE))
predicted <- within(predicted, {
  Predicted_Prob <- plogis(fit)
  LL <- plogis(fit - (1.96 * se.fit))
  UL <- plogis(fit + (1.96 * se.fit))
})

## view first few rows of final dataset
head(predicted)

ggplot(predicted, aes(x = open_water15, y = Predicted_Prob)) +
  geom_ribbon(aes(ymin = LL, ymax = UL), alpha = .2) +
  geom_line(size=1) + ylim(c(0,1))+ theme_no_legend


# TESTING THE LIKELIHOOD FUNCTION ON TERRESTRIAL LAND COVER

#likelihood function
#z = a vector of zeros and ones corresponding to available and used points
#cell = a vector of the cell numbers of each point
#surface = a data frame with the x/y coordinates and values of the raster surface

parms <- c(coef(lc_mods[["15"]])[1], coef(lc_mods[["15"]])[2], log(1))

NllBandwidth <- function(parms=parms, z=z, cell=cell, surface){
    #Set some things up:
    npx <- nrow(surface)
    b0 <- parms[1]              #intercept
    b1 <- parms[2]              #slope of the relationship
    b2 <- exp(parms[3])+1              #bw/spatial scale parameter

    r <- rasterFromXYZ(surface) #surface is a df with X|Y|VALUES
    zmat <- as.matrix(r)
    f <- kernel2dsmooth(zmat,kernel.type="gauss", nx=nrow(r), ny=ncol(r), sigma = b2)
    values(r) <- f
    ### Compute binomial success probabilities
    probs <- plogis(b0 + b1*(r[cell]))
    #ll <- rep(0,length(z))

    ### evaluate log of binomial pmf
    tmp <- dbinom(z,1,probs,log=TRUE)
    ll <-  -1*sum(tmp)
    return(ll)
  }


# Extract cell numbers
cell1 <- extract(open_water, ua_data[,c("x","y")], cellnumbers=TRUE)[,1]

# Convert raster to data frame
surface <- data.frame(coordinates(open_water), values(open_water))

# OPTIMIZE!!!!!!!!!!!!
system.time(out <- nlm(nll_kern_bw,parms,print.level=2,
                       z=ua_data$point,cell=cell1,
                       surface=surface,hessian=TRUE))




lik <- function(parms=params, z=z, cell=cell, surface=surface){

    #Set some things up:
    npx <- nrow(surface)
    b0 <- parms[1]              #intercept
    b1 <- parms[2]              #slope of the relationship
    b2 <- exp(parms[3])+1              #bw/spatial scale parameter

    r <- rasterFromXYZ(surface) #surface is a df with X|Y|VALUES
    zmat <- as.matrix(r)
    f <- kernel2dsmooth(zmat,kernel.type="gauss", nx=nrow(r), ny=ncol(r), sigma = b2)
    values(r) <- f
    ### Compute binomial success probabilities
    probs <- plogis(b0 + b1*(r[cell]))
    #ll <- rep(0,length(z))

    ### evaluate log of binomial pmf
    tmp <- dbinom(z,1,probs,log=TRUE)
    ll <-  -1*sum(tmp)
    return(ll)
}

# Extract cell numbers
cell1 <- extract(terrestrial,UAdata[,c("x","y")],cellnumbers=TRUE)[,1]

# Convert raster to data frame
surface <- data.frame(coordinates(terrestrial),values(terrestrial))

# OPTIMIZE!!!!!!!!!!!!
system.time(out <- nlm(lik,c(-1.301,3.620, log(67)),print.level=2,
                       z=UAdata$point,cell=cell1,
                       surface=surface,hessian=TRUE))


system.time(lik(parms,z,surface))


################
# This is code from Chris Sutherland's original script,
# including the simulation example.
################

par(mfrow=c(2,2))

#original layer
plot(terrestrial)


#smoothed layer
bw <- 900 / 15
zmat <- as.matrix(r.terrestrial)
zmat[zmat>0] <- 1
f <- kernel2dsmooth(zmat,kernel.type="gauss", nx=nrow(r.terrestrial), ny=ncol(r.terrestrial), sigma = bw)
r.terrestrial.smooth <- r.terrestrial
values(r.terrestrial.smooth) <- f
dev.new()
plot(r.terrestrial.smooth)



#binary surface
r.lp <- r.terrestrial.smooth
r.z <- r.terrestrial.smooth
lin.pred <- 3 + -30 * values(r.terrestrial.smooth)
values(r.lp) <- plogis(lin.pred)
plot(r.lp)
z <- rbinom(length(values(r.z)), 1, values(r.lp))
values(r.z) <- z
plot(r.z)

# cut out some points and limit the analysis to a subset of locations
surface <- data.frame(coordinates(r.terrestrial),values(r.terrestrial))
pick <- 1:nrow(surface)#sample(1:nrow(surface), 1000, replace=FALSE)
system.time(out <- nlm(lik,c(-0.89747,2.323, bw),z=z[pick],surface=surface[pick,],hessian=TRUE))





## Making used and available points --------------------------------------------

# Read in the used points and an equal number of available points

#load(file=file.path(getwd(), "used.RData"))
#load(file=file.path(getwd(), "avail.RData"))

all_nests <- read.csv(file=file.path("C:/Work/R/Data/BAEA/Nests",
  "Nests_Study_Intact_Last.csv"), header=TRUE, stringsAsFactors=FALSE)
all_nests <- ConvertNestIdToNum(all_nests)

study_nests <- read.csv(file=file.path("C:/Work/R/Data/BAEA/Nests/",
    "Nests_Study.csv"), header=TRUE, stringsAsFactors=FALSE) %>%
#  dplyr::filter(!name %in% c("Cherryfield", "Madagascal", "Webb", "Upper"))
  dplyr::filter(name %in% c("Sheepscot", "Sandy"))
study_nests <- ConvertNestIdToNum(study_nests)
study_nests[, "nest_id"] <- study_nests[, "nest_site"]

base <- raster("C:/ArcGIS/Data/BlankRaster/maine_30mc.tif")

baea_all <- read.csv("C:/Work/R/Data/BAEA/BAEA.csv", header=TRUE)
baea_sub <-  baea_all %>%
  dplyr::filter(agl <= 67) %>%
  dplyr::filter(speed <= 2) %>%
  dplyr::filter(year == 2015) %>%
#  dplyr::filter(!id %in% c("Cherryfield", "Madagascal", "Webb", "Upper"))
  dplyr::filter(id %in% c("Sheepscot", "Sandy"))

used <- baea_sub %>%
  mutate(x = long_utm) %>%
  mutate(y = lat_utm) %>%
  mutate(point = 1) %>%
  dplyr::select(x,y,point)
save(used, file = file.path(getwd(), "used.RData"))


source('C:/Work/R/Functions/gis.R')
ua_data <- SampleRandomPointsInHomerange(df_all=all_nests, df_home=study_nests,
  used_pts=baea_sub, base, max_r = 30000, id = "nest_site", name = "nest_id",
  output_dir = getwd(), write_homerange = TRUE)
#avail <-  avail1 %>%
#  mutate(x = long_utm) %>%
#  mutate(y = lat_utm) %>%
#  mutate(point = 0) %>%
#  dplyr::select(x,y,point)
save(ua_data, file = file.path(getwd(), "ua_data.RData"))
