==> R/3f_Analysis_Steps.R <==
#------------------------- Analysis Steps -------------------------------------#
# Create "baea_steps" from baea_movements
#------------------------------------------------------------------------------#

# Setup ------------------------------------------------------------------------

# Load packages
pacman::p_load(tidyverse)

## Create BAEA Movements -------------------------------------------------------

# Import baea and movement data
baea <- readRDS(file = "Data/BAEA/baea.rds")
baea_movements <- readRDS(file = "Data/BAEA/baea_movements.rds")

# Create locations
baea_locs <- baea %>%
  dplyr::select(date, datetime, id, long_utm, lat_utm, dx:step_time) %>%
  arrange(id, datetime) %>%
  group_by(id) %>%
  mutate(exp_angle = lag(abs_angle)) %>%
  mutate(long_utm_end = lead(long_utm)) %>%
  mutate(lat_utm_end = lead(lat_utm)) %>%
  ungroup() %>%
  dplyr::select(id, datetime, exp_angle, long_utm_end, lat_utm_end)

# Create steps
baea_steps <- left_join(baea_movements, baea_locs, by = c("id", "datetime")) %>%
  mutate(x = long_utm, y = lat_utm) %>%
  CenterXYWithBase(., base) %>%
  mutate(long_utm = x, lat_utm = y) %>%
  dplyr::select(id, datetime, behavior_behavior, long_utm, lat_utm, exp_angle,
    long_utm_end, lat_utm_end)

# Save files
saveRDS(baea_steps, file = "Data/BAEA/baea_steps.rds")
rm(baea, baea_steps, baea_locs)
